Split out shadow orders for future reuse
